pyproject.toml
[project]
name = "djangonewspaper"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita",email = "popovn2268@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-allauth[socialaccount] (>=65.4.1,<66.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
